pyproject.toml
[tool.poetry]
name = "rl-chat"
version = "0.1.0"
description = ""
authors = ["Demianight <komarovdemian@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.2"}
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
redis = "^5.1.1"
pyjwt = "^2.9.0"
celery = "^5.4.0"
aiogram = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
